rename file search_bar to searchBar
